DG_misc.h: Use size_t for DG_(v)snprintf()s size parameter
like C99 does.
It's consistent with other string functions taking size parameter, but
still weird, because (v)snprintf() are supposed to return the number of
bytes that would have been written with a big enough buffer - as an int.
Even on 32bit systems an int is not enough (size_t is unsigned int), and
and 64bit systems a (32bit) int definitely is not enough..
Thinking about asserting if someones passes a value > MAX_INT.

(I don't see a usecase for passing >2GB buffers to snprintf() anyway..)