SDL_stbimage.h: Add STBIMG_Load_RW_noSeek()
It doesn't use SDL_RWseek(), only SDL_RWsize() and SDL_RWread().
It uses that to read the whole file into a buffer and feeds that to
STBIMG_LoadFromMemory().
This is usefull for SDL_RWops streams that don't support seeking or are
slow at it, like on Android if SDL has to use Java InputStreams.