DG_misc.h: Add DG_memrmem() - like (DG_)memmem() but backwards + Tests
It searches for the *last* occurence of a byte-array's values in
another byte-array.

I've also written tests for this (because I couldn't reuse tests suites
from libc implementations or similar for this function), just compile
tests/testmemrmem.c and execute the resulting binary, it should print
"Success! All DG_memrmem() tests passed." to stdout and exit with status 0.