DG_misc.h: Rework DG_strnlen() to avoid reading past end of buffer
Until now DG_strnlen(str, len) used to read up to sizeof(uintptr_t)-1
bytes past str+len. While this *should* be safe (esp. when staying in
the same aligned machine word as the last valid byte) and is frequently
used in similar implementations of strlen(), but ASan hates it.
So I rewrote the whole thing and also optimized it as best as I could.
It's faster than before and doesn't read past the buffer anymore.

However, for DG_strlen() this means that by default it can only used a
very naive (and slow) implementation that checks one byte at a time,
because there we don't know the end of the buffer in advance.
With #define DG_MISC_STRLEN_OVERREAD you can make DG_strlen() use the
old trick again (Note that now in most cases even on Windows DG_strlen()
is just a #define to regular strlen() so this might not matter for you)

You can also #define DG_MISC_STRLEN_UNALIGNED to make DG_strnlen() use
a slightly different approach that reads the uintptr_t-sized chunks
unaligned. On some machines this is faster, on others slower.